nhw 메세지 개선

+++ core/scripts/nhw.task-iso.sh
@@ -4,18 +4,23 @@ run_iso_task() {
     log_msg "Task" "starting ISO image build process..."
     cd "$TMP_BUILD_DIR" || exit 1
     
-    log_exec "nom" ">"
-    nom build ".#nixosConfigurations.custom-iso.config.system.build.isoImage" \
-      --extra-experimental-features "nix-command flakes" --impure --print-build-logs
-    log_exec "nom" "<"
+    log_exec "nom" ">" "nom build iso"
+    if nom build ".#nixosConfigurations.custom-iso.config.system.build.isoImage" \
+      --extra-experimental-features "nix-command flakes" --impure --print-build-logs; then
+        log_exec "nom" "<" "nom build iso"
 
-    if [ -L "$TMP_BUILD_DIR/result" ]; then
-        ISO_FILE=$(readlink -f "$TMP_BUILD_DIR/result/iso/"*.iso)
-        ISO_NAME=$(basename "$ISO_FILE")
-        cp "$ISO_FILE" "$TMP_BUILD_DIR/"
-        rm -f "$TMP_BUILD_DIR/result"
-        log_msg "Done" "ISO successfully created: .build/$ISO_NAME"
+        if [ -L "$TMP_BUILD_DIR/result" ]; then
+            ISO_FILE=$(readlink -f "$TMP_BUILD_DIR/result/iso/"*.iso)
+            ISO_NAME=$(basename "$ISO_FILE")
+            cp "$ISO_FILE" "$TMP_BUILD_DIR/"
+            rm -f "$TMP_BUILD_DIR/result"
+            log_msg "Done" "ISO successfully created: .build/$ISO_NAME"
+        else
+            log_msg "Error" "ISO build failed: result link not found."
+            exit 1
+        fi
     else
+        log_exec "nom" "<" "nom build iso"
         log_msg "Error" "ISO build failed."
         exit 1
     fi
